+++ NEWS.md
@@ -2,8 +2,12 @@
 
 ## Major changes
 
-- First release
 - Start, stop and restart VMs
 - Create instances using cloudinit
 - Browser based SSH
+- SSH from R
 - Metadata
+- Docker
+- Google container registry
+- VM templates
+- Future asynch cluster computing